JCBC-581: Add a thinkTime to the integration tests when bucket is created.
Motivation
----------
The design doc tests hang when all are run in parallel. This is because a different
test removed all the buckets which is fine, but then the server can't flush the
new bucket so quickly.

The unsolved root cause is that we reliably don't know when the bucket is created.

Modifications
-------------
As a workaround, a configurable think time is added when the bucket is created. This is
most certainly only run once, since the bucket is not thrown away all the time.

Result
------
Integration tests now run through, even when intermingled.

Change-Id: I1b5b4c83cb393800713be4c6695fd1853b7760b1
Reviewed-on: http://review.couchbase.org/42074
Reviewed-by: Simon Baslé <simon.basle@gmail.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>