JCBC-584: Correctly forward CouchbaseExceptions and prevent from wrapping into Composite.
Motivation
----------
If an exception is re-thrown in an error handler, RxJava will wrap it into a CompositeException.
The intent was to just wrap it into a CouchbaseException if needed to provide a common base, and
the wrong handler method was used.

Modifications
-------------
The observable is now completed properly with the right sequence so that the Composite is not
added by RxJava and the toplevel Exception is still the CouchbaseException (or subclasses).

Result
------
Easier Exception-type based error handling during the connect process.

Change-Id: I677b7522567173fa047b78378ff831b54ec41333
Reviewed-on: http://review.couchbase.org/42111
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>