JCBC-585: Poll for all node on "healthy" state on bucket creation.
Motivation
----------
A better alternative to the previously implemented "think time" is to poll all nodes info
until they are in a "healthy" state - getting out of the "warmup" phases. While probably not
100% accurate, it should be good enough for testing and administrative scenarios.

Modifications
-------------
The code now polls every second for the bucket info, grabs the state of each node and only
proceeds if all of them are in a "healthy" state.

It also contains a small fix to the calculated bucket size because previously it did fail with
int/long conversions.

Result
------
Whena bucket is created or replaced, the code now waits until the bucket is in a healthy
state before reporting success.

Change-Id: Icbbf424682cbae04ad0f1fa83d4d1412851f08dc
Reviewed-on: http://review.couchbase.org/42117
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>