JCBC-590: Enable RawJsonTranscoder by default.
Motivation
----------
As part of the change introduced in JCBC-??, capabilities were added
to store and retrieve raw JSON data, but it was not enabled by default.

Modifications
-------------
Enables the RawJsonTranscoder (and its document type) by default, which allows
users to consume it without additional effort. Note that this should've been
in the older change and was an oversight.

Result
------
Proper use of the RawJsonDocument is now possible.

Change-Id: Ia5ddae858167cad13989a51837e49b091d5ad413
Reviewed-on: http://review.couchbase.org/42159
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>