Adjust checks in unit tests
Motivation
----------
Because of JsonDocument constuctor is not very verbose, it is very easy
to misuse it when it comes to setting expiry. Consider following two
signatures:

  JsonDocument create(String id, JsonObject content, long cas)
  JsonDocument create(String id, int expiry, JsonObject content)

Java does not warn when small literal number used.

The second point is that counter test clearly intent to assert that CAS
changed on each increment/decrement. But it done is some strange way.

Modifications
-------------
For first case, JsonDocument's expiry properly initialized. For the
second case, all combinations used to ensure that CAS does not match

Result
------
More robust unit tests, and more clear intention revealing.

Change-Id: I462a52cb50c1a89e3e7b5cce4037ea96c4fc50c6
Reviewed-on: http://review.couchbase.org/42331
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>