JCBC-601: Expose returned CAS value on remove() responses
Motivation
----------
Since remove calls on the server return the CAS value, it makes
sense to expose this information to the user.

Modifications
-------------
Builds on the core-io change (JVMCBC-48) and forwards the value into
the actual user-visible response. Test cases and documentation are
adapted appropriately.

Result
------
The user is able to consume the CAS value on remove.

Change-Id: I4369c1a6bb25d6f7fb5eeac4c4abd82378682fe6
Reviewed-on: http://review.couchbase.org/42346
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>