JCBC-602: Handle CAS mismatches on remove
Motivation
----------
When a CAS value is passed down together with a remove request,
the server performs a CAS check. Previously, the response was
not properly respected and the user had no idea if the check
succeeded or not.

Modifications
-------------
To be in line with the other methods, a CASMismatchException
is issued in case of such a state.

Result
------
The user can catch and handle such errors/exceptions now properly
and in line with similar methods.

Change-Id: I9dc619648227fe0294f42e36e9d3d4f7544b968e
Reviewed-on: http://review.couchbase.org/42347
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>