JCBC-600: Null out content() on append/prepend result.
Motivation
----------
When a content is appended/prepended to a already existing document,
the server protocol does not return the "final" content. It is better
to set the content to null instead of a wrong value - since the actual
content is not just the appended/prepended chunk.

Modifications
-------------
The content is nulled out on append and prepend. Tests have been added
and documentation improved.

Result
------
More accurate application semantics, as well as improved documentation
and testing.

Change-Id: I6648ef00525424f046d4dd2f1738a5581adb99f0
Reviewed-on: http://review.couchbase.org/42348
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>