JCBC-518: No intro page on the javadoc
Motivation
----------
We need a simple page to welcome the user and guide him to the best
place to get started. Also, external links to netty and other projects
are needed so that javadoc can resolve them.

Modifications
-------------
This change adds a welcome page with a simple pointer to the best
starting point, as well as external link references to the build
so they are picked up and the user can click his way through.

Result
------
Better navigateability through the javadocs.

Change-Id: I8af5513cabded82455b383ae441a11ddca28d451
Reviewed-on: http://review.couchbase.org/42765
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>