document IndexOutOfBoundsException, not null
Motivation
----------
In a JsonArray, the only error condition on get methods is when the
index provided is negative or too large. In this case, methods don't
return null as advertised in documentation but rather throw a
IndexOutOfBoundsException.
This is the correct behavior, and should be reflected in documentation.

Modifications
-------------
Documented the fact that get methods fail in IndexOutOfBoundsException
and not by returning null.
Added a quick unit test to show this Exception is expected.

Result
------
Documentation better reflects behavior of get methods, test passes.

Change-Id: Ib59f08c2caf15ac8b6d9edd8d513fa020c4549a1
Reviewed-on: http://review.couchbase.org/43362
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>