JCBC-576: Add support for geospatial views.
Motivation
----------
Geospatial indexing and querying will be a main feature of upcoming releases
and therefore needs to be supported in the java SDK. This implementation is
not compatible with the previous experimental geo support and can only be used
with subsequent server releases (see the product feature matrix for the actual
release version).

Modifications
-------------
Adding support consists of two items:

	- querying through the SpatialViewQuery
	- managing spatial views through the DesignDocument

A separate SpatialViewQuery has been introduced to reflect the builder that
only supports the params that a view query can actually understand. Responses
have been adapted but are very similar by intention to the regular view ones.

Result
------
Support for geospatial index querying in the new server versions. Not backwards
compatible with "older experimental" geo index implementations.

Change-Id: I1515710a7f2268c2e396b6adafb161903a59dbc4
Reviewed-on: http://review.couchbase.org/44015
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>