JCBC-653: Respect development flag on view query.
Motivation
----------
When development(boolean) is executed on the ViewQuery, the value
is ignored.

In addition, the BBOX param is not used in a regular view query,
so it should be removed.

Modifications
-------------
The changeset respects the development param and removes the unused
BBOX param from the internal datastructure.

Result
------
development(boolean) now works as expected and the unused BBOX param
is removed.

Change-Id: I16ff16cfdfd20e6897377c9c003011eee34c8a75
Reviewed-on: http://review.couchbase.org/44242
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-on: http://review.couchbase.org/44701