JCBC-657: Explicitly release content when not used in transcoder
Motivation
----------
When the content() is not explicitly in the transcoder, it needs to be
released so that on mutating operations, the pool does not leak. Note
that it will leak slowly and not really noticable, because the server
response is very small for non-get responses.

But it should be released so that the chunks can be reused by the pool.

Modifications
-------------
Explicitly release() the content buffer if not null and not used later.

Result
------
Better reusability for the pool on mutating operations, leads to less
memory pressure on the chunks.

Change-Id: I705141036b06a103b668ee395d3b426e32425ab3
Reviewed-on: http://review.couchbase.org/44380
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>