JCBC-661: View defer() needs to include the request on retry.
Motivation
----------
When defer() is used to retry an operation on resubscribe, the actual
request needs to be created fresh to, since it contains the Subject. If
the core detects that the request included subject is already completed,
it will not retry it but rather return immediately.

Modifications
-------------
Create a new request on every defer for both regular and spatial views.

Result
------
Correctly retrying the request and not failing downstream because
ByteBufs have already been used.

Change-Id: I4b55e35213fbfb52ea9b3490dcf05c1542b07532
Reviewed-on: http://review.couchbase.org/44474
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>