JCBC-662: Detect server feature availability
Motivation
----------
More and more features are only available on newer versions of server,
and the SDK has to maintain a clear picture of the compatibility matrix
which is getting harder to do explicitely via environment variables.

We need a way of easily and programatically detect if a feature is
available on a specific instance of a Cluster (eg Spatial View are only
found on cbserver >= 3.0.2).

Modifications
-------------
Added Version to represent a version number (major.minor.patch).
Added CouchbaseFeature, an enum of all testable features known to the
current version of the SDK with the server version from which each was
available.

Added helper class FeaturesHelper with a method to check for a feature
on a given Cluster (or rather the ClusterManager's ClusterInfo).
It will parse the info, find out the tiniest server version and check
that version vs the feature's minimum required version.

Added utility method 'ignoreIfMissing' on ClusterDependentTest to allow
integration tests to easily skip tests if the required feature is not
available.

Result
------
Anyone is now easily able to check if the connected cluster can provide
a given feature. Internally, integration tests can be skipped if that's
not the case.

All new tests pass.

Change-Id: Ia93cc63c8f383fc141896a5ff30a66c76d8b2445
Reviewed-on: http://review.couchbase.org/44520
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Simon Baslé <simon@couchbase.com>