JCBC-671: Ignore bad responses after retry handling.
Motivation
----------
When the view retry handler kicks in, it retries the view operation, but
the retryWhen handler emits all the "bad" responses as well. The code
afterwards instead only expects one good result to be emitted.

Modifications
-------------
The code now only returns the last() result, which is the good one, after
retrying succeeded. In addition, the code properly frees the info bytebufs,
since they are not accessed anymore when retried.

Result
------
Correct retry handling so that ever only one result is emitted.

Change-Id: I00cf65f49151e91b3db18122a59f399f0bd2fdf9
Reviewed-on: http://review.couchbase.org/44896
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-on: http://review.couchbase.org/44920
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Simon Baslé <simon@couchbase.com>