JCBC-673: Retry cancelled view requests too since they are idempotent.
Motivation
----------
View requests can be freely retried, even when the node is removed and
forefully cancelled because it was in-flight. This makes sure that
less view operations are cancelled at the user level.

Modifications
-------------
Explicitly also retry RequestCancelledExceptions from the core layer.

Result
------
Transparent view retries when a node is removed or failed over.

Change-Id: I71573232bd55b8a2f2deb28eb3807932586f8b1b
Reviewed-on: http://review.couchbase.org/44902
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-on: http://review.couchbase.org/44922
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Simon Baslé <simon@couchbase.com>