JCBC-675: Inverse logic for view retry on 404 responses.
Motivation
----------
The current code has troubles to decide if a 404 response should be
propagated or retried, since the response has been changed in 3.0.2
when a view is not found.

Modifications
-------------
This code inverts the logic, checking explicitly for the phrase returned
when a node is not provisioned, making sure only this case is retried
and the others are propagated.

Result
------
More robust handling of 404 responses against different node versions.

Change-Id: I0557050dff1242b782e05a3af1fe8346cfb917ec
Reviewed-on: http://review.couchbase.org/45021
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>