JCBC-682: Free ByteBufs on get* before filtering them out.
Motivation
----------
In all get* variants, the results are filtered on non-success. Since
content is still allocated in the responses, it needs to be properly
freed before filtering since the later code never has a chance to
do so properly.

Modifications
-------------
Free the allocated contents when the result is filtered out.

Result
------
No leaks when non-success responses for get* variants are returned.

Change-Id: I59a539e5fe6bb7cd0d9e54b1a5db1c9c93359bc3
Reviewed-on: http://review.couchbase.org/45310
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>