JCBC-674: Refactor prepared statements.
Motivation
----------
The current prepared statements API is great, but it can be made a bit
simpler in terms of working with statements. The API can hide the
PrepareStatement and therefore reduce the surface level for the user.

Modifications
-------------
queryPrepare has been streamlined to pepare and it now accepts any
statement. If it is a normal one, it is automatically prepared before
sent to the server.

Result
------
Easier way to access and work with prepared statements.

Change-Id: I2ac46c9b3c8575e1e5f00268b92fab4b873c31da
Reviewed-on: http://review.couchbase.org/45428
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>