JCBC-686: Make QueryResult iterable.
Motivation
----------
If the blocking query API is used, adding the Iterable interface provides
some nice syntactic sugar to iterate over rows with a for loop.

Modifications
-------------
Make the QueryResult implement the Iterable for the QueryRows and return
the iterator right away.

Result
------
Its now easier to just loop over the rows in the blocking api.

Change-Id: Icd167d84a16dc0d182203ec77b08f01520c25639
Reviewed-on: http://review.couchbase.org/45447
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>