Fix unit tests with > 3.0.0 compat
Motivation & Modifications
--------------------------
Spatial views have changed and currently it is not planned to support
the new format going forward on the 1.4 branch. It needs to be
conditionally disabled.

Also, the test config did incorrectly parse out the version of the nodes,
this has been corrected by backporting the Version regex from 2.1.

Change-Id: I43e9ad9ab2a48ec7fb5394564d380557cf869665
Reviewed-on: http://review.couchbase.org/45557
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>