JCBC-689: Throw exception when cluster manager passwd is wrong.
Motivation
----------
When a illegal password is used on the cluster manager, a proper
exception needs to be thrown that clearly shows what's going on.

Modifications
-------------
The code now checks the return status and throws a InvalidPasswordException
when it gets an unauthorized response from the server. If it is a different
error, a regular couchbase exception is thrown.

Result
------
Users will detect more quickly when they've entered a wrong password.

Change-Id: I8aefd1adae7ed93e8d15ac3cb35d6e7366e289b1
Reviewed-on: http://review.couchbase.org/45572
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>