JCBC-691: Expose n1ql's requestId/clientContextId
Motivation
----------
N1QL server assigns a requestId to each request and also lets the caller
set a clientContextId. Both are returned in the response, which allows
to track queries, log, logical operations, etc...

These should be exposed in the client's result.

Modifications
-------------
QueryResult and AsyncQueryResult now have getters for requestId and the
clientContextId.

Added tests for these getters in QueryTest + a test on truncation of the
clientContextId to 64 bytes on server side.

Improved QueryTests a bit.

Limitations
-----------
Current N1QL dp4 builds don't include consistency checking, which causes
first run of each statement to fail to return results... This has been
temporarily worked around by issuing a SELECT * in @BeforeClass.

Change-Id: Ibd7069884f043143c76f187ead8037b5af515e4c
Reviewed-on: http://review.couchbase.org/45860
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>