JCBC-542: Expose more specific KeyValue errors.
Motivation
----------
This changeset builds on top of JVMCBC-132 and exposes the specific error
codes as proper errors/exceptions to the user. This especially includes
E2BIG and TMPFAIL.

Modifications
-------------
The operations have been modified to take the new errors into account
and convert them into proper exceptions. In addition, a "catch all"
has been added.

Documentation for all new exceptions has been improved and added.

In addition to the return codes from the server, a proper check has
been put into place so that document IDs are not longer than 250
bytes - which would be rejected by the server anyway.

Result
------
Easier error handling and exposure for errors back up to the user
level.

Change-Id: Ic5262612156f0487186c0c2b23933caf22c5fc8b
Reviewed-on: http://review.couchbase.org/45874
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>